requirements.txt
coinbase-advanced-py
python-telegram-bot
python-dotenv
requests